refactor(tests): extract test_mem_chart_common, add gpu_arch coverage, repoint strip_ansi
